refacotr : Name 검증 책임 객체 구현
