+++ src/main/java/domain/player/Dealer.java
@@ -7,7 +7,7 @@ public class Dealer extends Player {
     private static final String DEALER_NAME = "딜러";
 
     public Dealer() {
-        super(DEALER_NAME);
+        super(new Name(DEALER_NAME));
     }
 
     public boolean canStand() {
