+++ src/main/java/domain/player/Name.java
@@ -1,0 +1,39 @@
+package domain.player;
+
+import expcetion.BlackjackException;
+import expcetion.ExceptionMessage;
+
+public class Name {
+
+    private static final int GAMBLER_NAME_MAX_LENGTH = 10;
+    private static final int GAMBLER_NAME_MIN_LENGTH = 2;
+    private static final String MATCH_NUMBER_PATTERN = ".*\\d.*";
+
+    private final String name;
+
+    public Name(String name) {
+        validate(name);
+        this.name = name;
+    }
+
+    private static void validate(String name) {
+        validateContainsNumber(name);
+        validateLength(name);
+    }
+
+    private static void validateContainsNumber(String name) {
+        if (name.matches(MATCH_NUMBER_PATTERN)) {
+            throw new BlackjackException(ExceptionMessage.NAME_NOT_INTEGER_ERROR);
+        }
+    }
+
+    private static void validateLength(String name) {
+        if (name.length() > GAMBLER_NAME_MAX_LENGTH || name.length() < GAMBLER_NAME_MIN_LENGTH) {
+            throw new BlackjackException(ExceptionMessage.NAME_LENGTH_ERROR);
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
